refactor: restructure project to follow Go conventions
