+++ internal/chatgpt/chatgpt.go
@@ -7,7 +7,7 @@ import (
 	openai "github.com/openai/openai-go/v3"
 	"github.com/openai/openai-go/v3/option"
 
-	"github.com/dfanso/commit-msg/src/types"
+	"github.com/dfanso/commit-msg/pkg/types"
 )
 
 func GenerateCommitMessage(config *types.Config, changes string, apiKey string) (string, error) {
